backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-otp==1.3.0  # Multi-factor authentication
qrcode==7.4.2  # For MFA QR codes
cryptography==42.0.0  # Encryption

# File Handling & Images
Pillow==10.2.0
pydicom==2.4.4
python-magic==0.4.27

# AWS
boto3==1.34.29  # S3, other AWS services

# Background Tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0  # Scheduled tasks
# API Documentation
drf-spectacular==0.27.1  # OpenAPI/Swagger docs

# Utilities
python-decouple==3.8  # Environment variables
django-extensions==3.2.3  # Shell plus, etc.
pytz==2024.1

# Medical/Healthcare Specific
hl7==0.4.0  # HL7 message parsing (updated package name)
fhir.resources==7.1.0  # FHIR support

# Data Processing
pandas==2.1.4  # For data migration/analysis
openpyxl==3.1.2  # Excel import/export

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0  # Test fixtures

# Code Quality
black==24.1.1  # Code formatting
flake8==7.0.0  # Linting
isort==5.13.2  # Import sorting

# Development
ipython==8.20.0
django-debug-toolbar==4.2.0
